% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation1.R
\docType{package}
\name{aggregation1}
\alias{aggregation1}
\alias{aggregation1-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
aggregation1
A data package for aggregation1.

Use \code{data(package='aggregation1')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
