% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation1.R
\docType{data}
\name{data}
\alias{data}
\title{Aggregation 1 data}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{subject}{}
\item{presentation}{}
\item{id}{}
\item{similarity}{}
\item{awareness}{}
\item{contact}{}
\item{address}{}
\item{prolific}{}
\item{sex}{}
\item{age}{}
\item{other}{}
\item{business_exp}{}
\item{business_edu}{}
\item{total_time}{}
\item{project}{}
\item{pos}{}
\item{dif}{}
\item{prob}{}
\item{project_order}{}
\item{choice}{}
\item{choice_aggregated}{}
\item{choice_neg_1}{}
\item{choice_neg_2}{}
\item{choice_samuelson_single}{}
\item{choice_samuelson_multiple}{}
\item{choice_samuelson_aggregated}{}
\item{choice_samuelson_original_single}{}
\item{choice_samuelson_original_multiple}{}
\item{choice_samuelson_original_aggregated}{}
\item{choice_redelmeier_single}{}
\item{choice_redelmeier_multiple}{}
\item{choice_redelmeier_aggregated}{}
\item{project_aggregated}{}
\item{project_neg_1}{}
\item{project_neg_2}{}
\item{project_samuelson_single}{}
\item{project_samuelson_multiple}{}
\item{project_samuelson_aggregated}{}
\item{project_samuelson_original_single}{}
\item{project_samuelson_original_multiple}{}
\item{project_samuelson_original_aggregated}{}
\item{project_redelmeier_single}{}
\item{project_redelmeier_multiple}{}
\item{project_redelmeier_aggregated}{}
\item{pos_aggregated}{}
\item{pos_neg_1}{}
\item{pos_neg_2}{}
\item{pos_samuelson_single}{}
\item{pos_samuelson_multiple}{}
\item{pos_samuelson_aggregated}{}
\item{pos_samuelson_original_single}{}
\item{pos_samuelson_original_multiple}{}
\item{pos_samuelson_original_aggregated}{}
\item{pos_redelmeier_single}{}
\item{pos_redelmeier_multiple}{}
\item{pos_redelmeier_aggregated}{}
\item{dif_aggregated}{}
\item{dif_neg_1}{}
\item{dif_neg_2}{}
\item{dif_samuelson_single}{}
\item{dif_samuelson_multiple}{}
\item{dif_samuelson_aggregated}{}
\item{dif_samuelson_original_single}{}
\item{dif_samuelson_original_multiple}{}
\item{dif_samuelson_original_aggregated}{}
\item{dif_redelmeier_single}{}
\item{dif_redelmeier_multiple}{}
\item{dif_redelmeier_aggregated}{}
\item{prob_aggregated}{}
\item{prob_neg_1}{}
\item{prob_neg_2}{}
\item{prob_samuelson_single}{}
\item{prob_samuelson_multiple}{}
\item{prob_samuelson_aggregated}{}
\item{prob_samuelson_original_single}{}
\item{prob_samuelson_original_multiple}{}
\item{prob_samuelson_original_aggregated}{}
\item{prob_redelmeier_single}{}
\item{prob_redelmeier_multiple}{}
\item{prob_redelmeier_aggregated}{}
\item{proportion}{}
\item{sample}{}
}
}
\source{
The data was collected using Prolific.
}
\description{
Detailed description of the data
}
\seealso{
\link{aggregation1}
}
