% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{extract_data}
\alias{extract_data}
\title{Extract data}
\usage{
extract_data(data_prep, names_to, names_pattern, values_to = "value")
}
\arguments{
\item{data_prep}{data}

\item{names_to}{pivot names}

\item{names_pattern}{pivot pattern}

\item{values_to}{pivot value}
}
\value{

}
\description{
For aggregation experiment
}
\examples{

\dontrun{extract_data(data_prep, names_to, names_pattern)}
}
