% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_name_materials_experiment1.R
\name{get_file_name_materials_experiment1}
\alias{get_file_name_materials_experiment1}
\title{Aggregation 1 screenshot file names}
\usage{
get_file_name_materials_experiment1()
}
\value{

}
\description{
Aggregation 1 screenshot file names
}
\author{
Shir Dekel
}
