% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omnibus.R
\name{get_omnibus}
\alias{get_omnibus}
\title{Get omnibus model for aggregation 1 experiment}
\usage{
get_omnibus(data)
}
\value{

}
\description{
Get omnibus model for aggregation 1 experiment
}
\author{
Shir Dekel
}
