% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_aggregation_1.R
\name{pivot_aggregation_1}
\alias{pivot_aggregation_1}
\title{Pivot aggregation 1 data

The goal is to keep all DVs in the same dataframe so that we can
subsequently just \code{nest_by()} to get what we want. Maybe it would have been
better to use separate dataframes.}
\usage{
pivot_aggregation_1(data)
}
\value{

}
\description{
Pivot aggregation 1 data

The goal is to keep all DVs in the same dataframe so that we can
subsequently just \code{nest_by()} to get what we want. Maybe it would have been
better to use separate dataframes.
}
\author{
Shir Dekel
}
