% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_JSON_tibble.R
\name{pull_JSON_tibble}
\alias{pull_JSON_tibble}
\title{Pull column, extract JSON, convert to tibble}
\usage{
pull_JSON_tibble(data, col)
}
\arguments{
\item{data}{Data}

\item{col}{Column to pull}
}
\value{

}
\description{
Pull column, extract JSON, convert to tibble
}
\author{
Shir Dekel
}
